requirements.txt
streamlit
azure-ai-documentintelligence
pandas
openpyxl
regex
python-Levenshtein
memory-profiler
psutil
